requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.40.0
greenlet==2.0.2
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
rawpy==0.18.1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
validators==0.21.2
virtualenv==20.24.3
Werkzeug==2.3.6
